FAT volume boot sector layout detection
